refactor(analysis): extract derived ratio owner
